requirements.txt
python-dotenv
langchain
replicate
llama-index
llama_index_llms_replicate
llama-index-embeddings.huggingface
